Core, Core.UI, Core.UI.WPF, Core.Windows - reorganize core libraries
