--- SoluiNet.DevTools.Core/Extensions/GridExtensions.cs
@@ -1,120 +1,0 @@
-﻿// <copyright file="GridExtensions.cs" company="SoluiNet">
-// Copyright (c) SoluiNet. All rights reserved.
-// </copyright>
-
-namespace SoluiNet.DevTools.Core.Extensions
-{
-    using System;
-    using System.Collections.Generic;
-    using System.Linq;
-    using System.Text;
-    using System.Threading.Tasks;
-    using System.Windows;
-    using System.Windows.Controls;
-    using System.Windows.Media;
-
-    /// <summary>
-    /// Provide a collection of methods which allow to work with grids.
-    /// </summary>
-    public static class GridExtensions
-    {
-        /// <summary>
-        /// Get a child of a grid which implements the overgiven type.
-        /// </summary>
-        /// <typeparam name="T">The type of the child element.</typeparam>
-        /// <param name="depObj">The grid which should be containing the child.</param>
-        /// <returns>Returns a child object which is contained in the overgiven grid and has the overgiven type.</returns>
-        public static T GetChildOfType<T>(this DependencyObject depObj)
-            where T : DependencyObject
-        {
-            if (depObj == null)
-            {
-                return null;
-            }
-
-            for (var i = 0; i < VisualTreeHelper.GetChildrenCount(depObj); i++)
-            {
-                var child = VisualTreeHelper.GetChild(depObj, i);
-
-                var result = (child as T) ?? GetChildOfType<T>(child);
-
-                if (result != null)
-                {
-                    return result;
-                }
-            }
-
-            return null;
-        }
-
-        /// <summary>
-        /// Finds a Child of a given item in the visual tree.
-        /// </summary>
-        /// <param name="parent">A direct parent of the queried item.</param>
-        /// <typeparam name="T">The type of the queried item.</typeparam>
-        /// <param name="childName">x:Name or Name of child. </param>
-        /// <returns>The first parent item that matches the submitted type parameter.
-        /// If not matching item can be found,
-        /// a null parent is being returned.</returns>
-        public static T FindChild<T>(this DependencyObject parent, string childName)
-            where T : DependencyObject
-        {
-            // Confirm parent and childName are valid.
-            if (parent == null)
-            {
-                return null;
-            }
-
-            T foundChild = null;
-
-            int childrenCount = VisualTreeHelper.GetChildrenCount(parent);
-            for (int i = 0; i < childrenCount; i++)
-            {
-                var child = VisualTreeHelper.GetChild(parent, i);
-
-                // If the child is not of the request child type child
-                T childType = child as T;
-                if (childType == null)
-                {
-                    // recursively drill down the tree
-                    foundChild = FindChild<T>(child, childName);
-
-                    // If the child is found, break so we do not overwrite the found child.
-                    if (foundChild != null)
-                    {
-                        break;
-                    }
-                }
-                else if (!string.IsNullOrEmpty(childName))
-                {
-                    var frameworkElement = child as FrameworkElement;
-
-                    // If the child's name is set for search
-                    if (frameworkElement != null && frameworkElement.Name == childName)
-                    {
-                        // if the child's name is of the request name
-                        foundChild = (T)child;
-                        break;
-                    }
-
-                    // recursively drill down the tree
-                    foundChild = FindChild<T>(child, childName);
-
-                    // If the child is found, break so we do not overwrite the found child.
-                    if (foundChild != null)
-                    {
-                        break;
-                    }
-                }
-                else
-                {
-                    // child element found.
-                    foundChild = (T)child;
-                    break;
-                }
-            }
-
-            return foundChild;
-        }
-    }
-}
