+++ SoluiNet.DevTools.UI/MainWindow.xaml.cs
@@ -31,10 +31,10 @@ namespace SoluiNet.DevTools.UI
     using SoluiNet.DevTools.Core.ScriptEngine;
     using SoluiNet.DevTools.Core.Tools;
     using SoluiNet.DevTools.Core.Tools.Sql;
-    using SoluiNet.DevTools.Core.Tools.UI;
     using SoluiNet.DevTools.Core.UI;
     using SoluiNet.DevTools.Core.UI.Application;
     using SoluiNet.DevTools.Core.UI.WPF.Application;
+    using SoluiNet.DevTools.Core.UI.WPF.Tools.UI;
     using SoluiNet.DevTools.Core.UI.WPF.Window;
 
     /// <summary>
