+++ SoluiNet.DevTools.Utils.TimeTracking/Entities/TimeTrackingContext.cs
@@ -14,8 +14,8 @@ namespace SoluiNet.DevTools.Utils.TimeTracking.Entities
     using System.Text.RegularExpressions;
     using System.Windows.Media;
     using NLog;
-    using SoluiNet.DevTools.Core.Tools.UI;
     using SoluiNet.DevTools.Core.Tools.XML;
+    using SoluiNet.DevTools.Core.UI.WPF.Tools.UI;
     using SoluiNet.DevTools.Core.XmlData;
 
     /// <summary>
