+++ SoluiNet.DevTools.sln
@@ -65,9 +65,11 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Install", "Install", "{F99C
 EndProject
 Project("{930C7802-8A8C-48F9-8165-68863BCCD9DD}") = "SoluiNet.DevTools.Install", "SoluiNet.DevTools.Install\SoluiNet.DevTools.Install.wixproj", "{040B16E4-BEC3-4429-920D-C0D206B1D6D9}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SoluiNet.DevTools.Core.UI", "SoluiNet.DevTools.Core.UI\SoluiNet.DevTools.Core.UI.csproj", "{01B12B16-26AB-41EF-9D2F-91894F0CE19A}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SoluiNet.DevTools.Core.UI", "SoluiNet.DevTools.Core.UI\SoluiNet.DevTools.Core.UI.csproj", "{01B12B16-26AB-41EF-9D2F-91894F0CE19A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SoluiNet.DevTools.Core.UI.WPF", "SoluiNet.DevTools.Core.UI.WPF\SoluiNet.DevTools.Core.UI.WPF.csproj", "{E7AE24F6-004C-4CE3-86D5-58D7F5FE89CC}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SoluiNet.DevTools.Core.UI.WPF", "SoluiNet.DevTools.Core.UI.WPF\SoluiNet.DevTools.Core.UI.WPF.csproj", "{E7AE24F6-004C-4CE3-86D5-58D7F5FE89CC}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SoluiNet.DevTools.Core.Windows", "SoluiNet.DevTools.Core.Windows\SoluiNet.DevTools.Core.Windows.csproj", "{64B47EED-3674-47F1-80FE-969339A8D34C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -373,6 +375,18 @@ Global
 		{E7AE24F6-004C-4CE3-86D5-58D7F5FE89CC}.Release|x64.Build.0 = Release|Any CPU
 		{E7AE24F6-004C-4CE3-86D5-58D7F5FE89CC}.Release|x86.ActiveCfg = Release|Any CPU
 		{E7AE24F6-004C-4CE3-86D5-58D7F5FE89CC}.Release|x86.Build.0 = Release|Any CPU
+		{64B47EED-3674-47F1-80FE-969339A8D34C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{64B47EED-3674-47F1-80FE-969339A8D34C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{64B47EED-3674-47F1-80FE-969339A8D34C}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{64B47EED-3674-47F1-80FE-969339A8D34C}.Debug|x64.Build.0 = Debug|Any CPU
+		{64B47EED-3674-47F1-80FE-969339A8D34C}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{64B47EED-3674-47F1-80FE-969339A8D34C}.Debug|x86.Build.0 = Debug|Any CPU
+		{64B47EED-3674-47F1-80FE-969339A8D34C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{64B47EED-3674-47F1-80FE-969339A8D34C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{64B47EED-3674-47F1-80FE-969339A8D34C}.Release|x64.ActiveCfg = Release|Any CPU
+		{64B47EED-3674-47F1-80FE-969339A8D34C}.Release|x64.Build.0 = Release|Any CPU
+		{64B47EED-3674-47F1-80FE-969339A8D34C}.Release|x86.ActiveCfg = Release|Any CPU
+		{64B47EED-3674-47F1-80FE-969339A8D34C}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -406,6 +420,7 @@ Global
 		{040B16E4-BEC3-4429-920D-C0D206B1D6D9} = {F99C02DD-E20B-4564-85FE-C6DB5BCF57A6}
 		{01B12B16-26AB-41EF-9D2F-91894F0CE19A} = {A8A6170C-6952-4985-8317-D765DB05FE47}
 		{E7AE24F6-004C-4CE3-86D5-58D7F5FE89CC} = {A8A6170C-6952-4985-8317-D765DB05FE47}
+		{64B47EED-3674-47F1-80FE-969339A8D34C} = {A8A6170C-6952-4985-8317-D765DB05FE47}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {8F61C6A5-176D-49B1-99D0-6BDBD8E0168A}
